Fixing the wrong call of this._pingPopup.close()
This in the call back function of observable, which I added in the last PR, is not the right this object. 
And we should keep the ping tab opened for a while according to the below comment from Jimmy.

```
The popup allows users to receive the browser prompt to allow requests to the IIS Administration API which uses an untrusted certificate. The certificate may be untrusted when accessing from a computer other than the one that the API was installed on. AJAX requests to an untrusted certificate are simply blocked, but opening up the popup allows the user to bypass it with the in-browser prompt.

The tab shouldn’t be closed before successfully pinging the API, otherwise it will just be closed before the user can click through the untrusted certificate prompt.

One option to close the tab gracefully is to update the IIS Administration API so that the page that is loaded in the popup automatically closes itself. Or we could display a prompt to the user

“Hello, this page is used to ensure that a connection can be established. Everything is in order. Please “click here” to close!”
```